setup for initial deployment
