+++ The_Site/Flask-Files/Just-Search-Fly-Deployable/Procfile
@@ -1,0 +1,2 @@
+# TODO: Modify this Procfile to fit your needs
+web: gunicorn justSearch:app
