+++ The_Site/Flask-Files/Just-Search-Fly-Deployable/fly.toml
@@ -1,0 +1,21 @@
+# fly.toml app configuration file generated for search-engine-api on 2023-09-17T17:43:03-07:00
+#
+# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
+#
+
+app = "search-engine-api"
+primary_region = "phx"
+
+[build]
+  builder = "paketobuildpacks/builder:base"
+
+[env]
+  PORT = "8080"
+
+[http_service]
+  internal_port = 8080
+  force_https = true
+  auto_stop_machines = true
+  auto_start_machines = true
+  min_machines_running = 0
+  processes = ["app"]
