+++ The_Site/Flask-Files/Just-Search-Fly-Deployable/requirements.txt
@@ -1,0 +1,7 @@
+Flask==2.3.2
+Flask_Cors==3.0.10
+flask_sqlalchemy==3.1.1
+nltk==3.8.1
+numpy==1.24.2
+SQLAlchemy==2.0.12
+gunicorn==21.2.0
